requirements.txt
# Core Dependencies
torch>=1.13.0
transformers>=4.21.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Deep Learning
pytorch-lightning>=1.8.0
tensorboard>=2.11.0
wandb>=0.13.0

# Android Analysis
androguard>=3.4.0
lxml>=4.9.0

# Frida (Dynamic Instrumentation)
frida>=16.0.0
frida-tools>=12.0.0

# Graph Processing
networkx>=2.8.0
pydot>=1.4.2
graphviz>=0.20.0

# Data Processing
jsonlines>=3.1.0
pyyaml>=6.0
tqdm>=4.64.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Database
psycopg2-binary>=2.9.0
redis>=4.4.0
sqlalchemy>=1.4.0

# API & Web
fastapi>=0.88.0
uvicorn>=0.20.0
requests>=2.28.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Code Quality
black>=22.10.0
flake8>=6.0.0
mypy>=0.991

# Utilities
python-dotenv>=0.21.0
colorlog>=6.7.0
joblib>=1.2.0

# NLP & Tokenization
tokenizers>=0.13.0
sentencepiece>=0.1.97

# Statistical Analysis
scipy>=1.9.0
statsmodels>=0.13.0

# ML Experiment Tracking
mlflow>=2.0.0

# APK Processing (Python wrappers)
# Note: Soot and Jadx require Java installation
